+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/dynamodbattributes_test.go
@@ -31,7 +31,7 @@ func TestDynamodbTagsBatchGetItemInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice("aws.dynamodb.table_names", []string{"table1"}))
 }
@@ -60,7 +60,7 @@ func TestDynamodbTagsBatchWriteItemInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice("aws.dynamodb.table_names", []string{"table1"}))
 }
@@ -114,7 +114,7 @@ func TestDynamodbTagsCreateTableInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -144,7 +144,7 @@ func TestDynamodbTagsDeleteItemInput(t *testing.T) {
 			TableName: aws.String("table1"),
 		},
 	}
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -157,7 +157,7 @@ func TestDynamodbTagsDeleteTableInput(t *testing.T) {
 			TableName: aws.String("table1"),
 		},
 	}
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -170,7 +170,7 @@ func TestDynamodbTagsDescribeTableInput(t *testing.T) {
 			TableName: aws.String("table1"),
 		},
 	}
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -184,7 +184,7 @@ func TestDynamodbTagsListTablesInput(t *testing.T) {
 			Limit:                   aws.Int32(10),
 		},
 	}
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.String("aws.dynamodb.exclusive_start_table", "table1"))
 	assert.Contains(t, attributes, attribute.Int("aws.dynamodb.limit", 10))
@@ -202,7 +202,7 @@ func TestDynamodbTagsPutItemInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -230,7 +230,7 @@ func TestDynamodbTagsQueryInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"table1"},
@@ -257,7 +257,7 @@ func TestDynamodbTagsScanInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"my-table"},
@@ -285,7 +285,7 @@ func TestDynamodbTagsUpdateItemInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"my-table"},
@@ -326,7 +326,7 @@ func TestDynamodbTagsUpdateTableInput(t *testing.T) {
 		},
 	}
 
-	attributes := DynamoDBAttributeSetter(context.TODO(), input)
+	attributes := DynamoDBAttributeBuilder(context.TODO(), input, middleware.InitializeOutput{})
 
 	assert.Contains(t, attributes, attribute.StringSlice(
 		"aws.dynamodb.table_names", []string{"my-table"},