+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/snsattributes.go
@@ -15,7 +15,13 @@ import (
 )
 
 // SNSAttributeSetter sets SNS specific attributes depending on the SNS operation is being performed.
+// Deprecated: Kept for backward compatibility, use SNSAttributeBuilder instead. This will be removed in a future release.
 func SNSAttributeSetter(ctx context.Context, in middleware.InitializeInput) []attribute.KeyValue {
+	return SNSAttributeBuilder(ctx, in, middleware.InitializeOutput{})
+}
+
+// SNSAttributeBuilder sets SNS specific attributes depending on the SNS operation is being performed.
+func SNSAttributeBuilder(ctx context.Context, in middleware.InitializeInput, out middleware.InitializeOutput) []attribute.KeyValue {
 	snsAttributes := []attribute.KeyValue{semconv.MessagingSystemKey.String("aws_sns")}
 
 	switch v := in.Parameters.(type) {